+++ tests/test_enterprise/api/test_enterprise_customer_admin.py
@@ -1,10 +1,11 @@
 """
-Tests for the EnterpriseCustomerAdminViewSet.
+Tests for the `EnterpriseCustomerAdminViewSet`.
 """
 
-from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
+
+from django.urls import reverse
 
 from enterprise.models import EnterpriseCustomerAdmin
 from test_utils.factories import (