+++ CHANGELOG.rst
@@ -17,9 +17,17 @@ Unreleased
 ----------
 * nothing unreleased
 
-[5.13.2]
+[5.13.4]
 ---------
 * feat: adds EnterpriseCustomerAdminViewSet
+
+[5.13.3]
+--------
+* fix: fixed chat_completion endpoint to handle Xpert AI v2 response correctly.
+
+[5.13.2]
+--------
+* fix: create migration for unmanage django model that points at DB view.
 
 [5.13.1]
 --------